requirements.txt
pandas
scikit-learn
streamlit
langchain
chromadb
sentence-transformers
matplotlib
python-dotenv
requests
openai
mistralai                          